organize server
