+++ server/src/classes/moduleOptions.interface.ts
