+++ server/src/modules/metadataModule/metadataOptions.interface.ts
@@ -1,4 +1,4 @@
-import { ModuleOptions } from '../../classes/moduleOptions';
+import { ModuleOptions } from '../../classes/moduleOptions.interface';
 
 export default interface MetadataOptions extends ModuleOptions {
 	dateAfter?: Date;
